pyproject.toml
[project]
name = "onlocation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tzdata>=2025.3",
]
